+++ README.md
@@ -64,6 +64,7 @@ RCON_PORT=xx # The rcon port
 RCON_PASSWORD=xx # The rcon password
 FACTORIO_LOG=C:\Users\xx\AppData\Roaming\Factorio\console.log (Unix path also supported) # Path to the chat log (--console-log)
 MOD_LOG=C:\Users\xx\AppData\Roaming\Factorio\script-output\factorigo-chat-bot\factorigo-chat-bot.log (Optional) # If you use the companion mod supply the path to that log file here
+POLL_LOG=1 # If this is set, logs will be polled instead of using inotify. This may be required with docker depending on your storage setup
 ```
 When using docker: don't forget to also mount/bind the log-files to your container.
