+++ go.mod
@@ -5,7 +5,6 @@ go 1.17
 require (
 	github.com/bwmarrin/discordgo v0.26.1
 	github.com/forewing/csgo-rcon v1.3.0
-	github.com/fsnotify/fsnotify v1.5.1
 	github.com/joho/godotenv v1.4.0
 	github.com/sirupsen/logrus v1.8.1
 )
@@ -13,7 +12,11 @@ require (
 require (
 	github.com/forPelevin/gomoji v1.1.2 // indirect
 	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/hpcloud/tail v1.0.0 // indirect
+	github.com/nxadm/tail v1.0.0 // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
 	golang.org/x/crypto v0.4.0 // indirect
 	golang.org/x/sys v0.3.0 // indirect
+	gopkg.in/fsnotify.v1 v1.4.7 // indirect
+	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
 )
